requirements.txt
python-telegram-bot==13.15
feedparser==6.0.12
python-dotenv==1.0.1
urllib3==1.26.18
